refactor: extract card title into separate component for reuseable
