+++ packages/page-governance/src/components/CardTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './CardTitle';
