+++ packages/page-governance/src/components/CardTitle/styled/index.ts
@@ -1,0 +1,1 @@
+export * from './TextWrapper.styled';
