+++ packages/page-governance/src/components/ProposalCard/styled/index.ts
@@ -1,4 +1,3 @@
 export * from './BadgesGroup.styled';
 export * from './BadgeWrapper.styled';
 export * from './ProposalCardWrapper.styled';
-export * from './TextWrapper.styled';
